+++ pypeit/coadd2d.py
@@ -49,6 +49,7 @@ class CoAdd2D:
     # Superclass factory method generates the subclass instance
     @classmethod
     def get_instance(cls, spec2dfiles, spectrograph, par, det=1, offsets=None, weights='auto',
+                     only_slits=None, exclude_slits=None,
                      spec_samp_fact=1.0, spat_samp_fact=1.0,
                      sn_smooth_npix=None, bkg_redux=False, find_negative=False, show=False,
                      show_peaks=False, debug_offsets=False, debug=False):
@@ -65,15 +66,16 @@ class CoAdd2D:
             :class:`CoAdd2D` as its base.
         """
 
-        return next(c for c in cls.__subclasses__() 
+        return next(c for c in cls.__subclasses__()
                     if c.__name__ == (spectrograph.pypeline + 'CoAdd2D'))(
                         spec2dfiles, spectrograph, par, det=det, offsets=offsets, weights=weights,
+                        only_slits=only_slits, exclude_slits=exclude_slits,
                         spec_samp_fact=spec_samp_fact, spat_samp_fact=spat_samp_fact,
                         sn_smooth_npix=sn_smooth_npix, bkg_redux=bkg_redux, find_negative=find_negative,
                         show=show, show_peaks=show_peaks, debug_offsets=debug_offsets, debug=debug)
 
     def __init__(self, spec2d, spectrograph, par, det=1, offsets=None, weights='auto',
-                 spec_samp_fact=1.0, spat_samp_fact=1.0,
+                 only_slits=None, exclude_slits=None, spec_samp_fact=1.0, spat_samp_fact=1.0,
                  sn_smooth_npix=None, bkg_redux=False, find_negative=False, show=False,
                  show_peaks=False, debug_offsets=False, debug=False):
         """
@@ -91,7 +93,7 @@ class CoAdd2D:
                 must include detectors viable as a mosaic for the provided
                 spectrograph; see
                 :func:`~pypeit.spectrographs.spectrograph.Spectrograph.allowed_mosaics`.
-            offsets (`numpy.ndarray`_, optional):
+            offsets (`numpy.ndarray`_ or string, optional):
                 Spatial offsets to be applied to each image before coadding. Here are the possible usage
                 cases:
 
@@ -110,6 +112,10 @@ class CoAdd2D:
                                     otherwise use uniform weights
                      3) list or array = Input a list of len(nexp) or shape = (nexp,) and
                                         the routine will use the input weights
+            only_slits (:obj:`list`, optional):
+                List of slits to coadd. It must be `slitord_id`.
+            exclude_slits (:obj:`list`, optional):
+                List of slits to exclude from coaddition. It must be `slitord_id`.
             spec_samp_fact (:obj:`float`, optional):
                 Make the wavelength grid sampling finer (``spec_samp_fact`` less
                 than 1.0) or coarser (``spec_samp_fact`` greater than 1.0) by
@@ -197,14 +203,11 @@ class CoAdd2D:
         self.good_slits = None
         self.maskdef_offset = None
 
-
         # Load the stack_dict
         self.stack_dict = self.load_coadd2d_stacks(self.spec2d)
         self.pypeline = self.spectrograph.pypeline
 
         self.nexp = len(self.spec2d)
-
-        self.exptime_coadd = self.stack_dict['exptime_coadd']
 
         # Check that there are the same number of slits on every exposure
         nslits_list = [slits.nslits for slits in self.stack_dict['slits_list']]
@@ -232,8 +235,18 @@ class CoAdd2D:
         # If smoothing is not input, smooth by 10% of the maximum spectral dimension
         self.sn_smooth_npix = sn_smooth_npix if sn_smooth_npix is not None else 0.1*self.nspec_max
 
+        # coadded frame parameters
+
+        # get slit index that indicates which slits are good for coadding
+        self.good_slits = self.good_slitindx(only_slits=only_slits, exclude_slits=exclude_slits)
+        # get the number of slits that are going to be coadded
+        self.nslits_coadded = self.good_slits.size
+
+        # effective exposure time
+        self.exptime_coadd = self.stack_dict['exptime_coadd']
+
     @staticmethod
-    def default_par(spectrograph, inp_cfg=None, det=None, slits=None):
+    def default_par(spectrograph, inp_cfg=None, det=None, only_slits=None, exclude_slits=None):
         """
         Get the default 2D coadding parameters.
 
@@ -245,8 +258,12 @@ class CoAdd2D:
                 An existing set of parameters to add to.
             det (:obj:`list`, :obj:`str`, :obj:`tuple`, optional):
                 Limit the coadding to this (set of) detector(s)/detector mosaic(s)
-            slits (:obj:`list`, :obj:`str`, optional):
-                Limit the coadding to this (set of) slit(s)
+            only_slits (:obj:`list`, :obj:`str`, optional):
+                Limit the coadding to this (set of) slit(s). Only_slits and exclude_slits
+                are mutually exclusive. If both are set, only_slits takes precedence.
+            exclude_slits (:obj:`list`, :obj:`str`, optional):
+                Exclude this (set of) slit(s) from the coadding. Only_slits and exclude_slits
+                are mutually exclusive. If both are set, only_slits takes precedence.
 
         Returns:
             :obj:`dict`: The default set of parameters.
@@ -254,11 +271,27 @@ class CoAdd2D:
         cfg = dict(rdx=dict(spectrograph=spectrograph))
         if inp_cfg is not None:
             cfg = utils.recursive_update(cfg, dict(inp_cfg))
+        if only_slits is not None and det is not None:
+            msgs.warn('only_slits and det are mutually exclusive. Ignoring det.')
+            _det = None
+        else:
+            _det = det
+
         if det is not None:
-            cfg['rdx']['detnum'] = det
-        if slits is not None:
+            cfg['rdx']['detnum'] = _det
+
+        if only_slits is not None and exclude_slits is not None:
+            msgs.warn('only_slits and exclude_slits are mutually exclusive. Ignoring exclude_slits.')
+            _exclude_slits = None
+        else:
+            _exclude_slits = exclude_slits
+
+        if only_slits is not None:
             utils.add_sub_dict(cfg, 'coadd2d')
-            cfg['coadd2d']['only_slits'] = slits
+            cfg['coadd2d']['only_slits'] = only_slits
+        if _exclude_slits is not None:
+            utils.add_sub_dict(cfg, 'coadd2d')
+            cfg['coadd2d']['exclude_slits'] = _exclude_slits
         # TODO: Heliocentric for coadd2d needs to be thought through. Currently
         # turning it off.
         utils.add_sub_dict(cfg, 'calibrations')
@@ -304,7 +337,7 @@ class CoAdd2D:
                 f"{io.remove_suffix(lasthdr['FILENAME'])}-{frsthdr['TARGET']}"
 
     @staticmethod
-    def output_paths(spec2d_files, par):
+    def output_paths(spec2d_files, par, coadd_dir=None):
         """
         Construct the names and ensure the existence of the science and QA output directories.
 
@@ -321,6 +354,9 @@ class CoAdd2D:
                 Full set of parameters.  The only used parameters are
                 ``par['rdx']['scidir']`` and ``par['rdx']['qadir']``.  WARNING:
                 This also *alters* the value of ``par['rdx']['qadir']``!!
+            coadd_dir (:obj:`str`, optional):
+                Path to the directory to use for the coadd2d output.
+                If None, the parent of the science directory is used.
 
         Returns:
             :obj:`tuple`: Two strings with the names of (1) the science output
@@ -328,7 +364,10 @@ class CoAdd2D:
             creates both directories if they do not exist.
         """
         # Science output directory
-        pypeit_scidir = Path(spec2d_files[0]).parent
+        if coadd_dir is not None:
+            pypeit_scidir = Path(coadd_dir).resolve() / 'Science'
+        else:
+            pypeit_scidir = Path(spec2d_files[0]).parent
         coadd_scidir = pypeit_scidir.parent / f"{par['rdx']['scidir']}_coadd"
         if not coadd_scidir.exists():
             coadd_scidir.mkdir(parents=True)
@@ -339,7 +378,7 @@ class CoAdd2D:
             qa_path.mkdir(parents=True)
         return str(coadd_scidir), str(qa_path)
 
-    def good_slitindx(self, only_slits=None):
+    def good_slitindx(self, only_slits=None, exclude_slits=None):
         """
         This provides an array of index of slits in the un-coadded frames that are considered good for 2d coadding.
         A bitmask common to all the un-coadded frames is used to determine which slits are good. Also,
@@ -347,14 +386,24 @@ class CoAdd2D:
 
         Args:
             only_slits (:obj:`list`, optional):
-                List of slits to combine. It must be `slitord_id`
+                List of slits to combine. It must be `slitord_id`.
+                Only_slits and exclude_slits are mutually exclusive.
+                If both are provided, only_slits takes precedence.
+            exclude_slits (:obj:`list`, optional):
+                List of slits to exclude. It must be `slitord_id`.
+                Only_slits and exclude_slits are mutually exclusive.
+                If both are provided, only_slits takes precedence.
 
         Returns:
             `numpy.ndarray`_: array of index of good slits in the un-coadded frames
         """
 
-        only_slits = [only_slits] if (only_slits is not None and
-                                      isinstance(only_slits, (int, np.integer))) else only_slits
+        if exclude_slits is not None and only_slits is not None:
+            msgs.warn('Both `only_slits` and `exclude_slits` are provided. They are mutually exclusive. '
+                      'Using `only_slits` and ignoring `exclude_slits`')
+            _exclude_slits = None
+        else:
+            _exclude_slits = exclude_slits
 
         # This creates a unified bpm common to all frames
         slits0 = self.stack_dict['slits_list'][0]
@@ -366,26 +415,43 @@ class CoAdd2D:
             slits = self.stack_dict['slits_list'][i]
             reduce_bpm |= (slits.mask > 0) & (np.invert(slits.bitmask.flagged(slits.mask,
                                                                               flag=slits.bitmask.exclude_for_reducing)))
-        # this are the good slit index according to the bpm mask
+        # these are the good slit index according to the bpm mask
         good_slitindx = np.where(np.logical_not(reduce_bpm))[0]
 
         # If we want to coadd all the good slits
-        if only_slits is None:
+        if only_slits is None and _exclude_slits is None:
             return good_slitindx
 
         # If instead we want to coadd only a selected (by the user) number of slits
-        # this are the `slitord_id` of the slits that we want to coadd
-        only_slits = np.atleast_1d(only_slits)
-        # create an array of slit index that are selected by the user and are also good slits
-        good_onlyslits = np.array([], dtype=int)
-        for islit in only_slits:
-            if islit not in slits0.slitord_id[good_slitindx]:
-                # Warnings for the slits that are selected by the user but NOT good slits
-                msgs.warn('Slit {} cannot be coadd because masked'.format(islit))
-            else:
-                indx = np.where(slits0.slitord_id[good_slitindx] == islit)[0]
-                good_onlyslits = np.append(good_onlyslits, good_slitindx[indx])
-        return good_onlyslits
+        if only_slits is not None:
+            # these are the `slitord_id` of the slits that we want to coadd
+            _only_slits = np.atleast_1d(only_slits)
+            # create an array of slit index that are selected by the user and are also good slits
+            good_onlyslits = np.array([], dtype=int)
+            msgs.info('Coadding only the following slits:')
+            for islit in _only_slits:
+                if islit not in slits0.slitord_id[good_slitindx]:
+                    # Warnings for the slits that are selected by the user but NOT good slits
+                    msgs.warn('Slit {} cannot be coadd because masked'.format(islit))
+                else:
+                    msgs.info(f'Slit {islit}')
+                    indx = np.where(slits0.slitord_id[good_slitindx] == islit)[0]
+                    good_onlyslits = np.append(good_onlyslits, good_slitindx[indx])
+            return good_onlyslits
+
+        # if we want to exclude some slits (selected by the user) from coadding
+        # these are the `slitord_id` of the slits that we want to exclude
+        _exclude_slits = np.atleast_1d(_exclude_slits)
+        # create an array of slit index that are excluded by the user
+        exclude_slitindx = np.array([], dtype=int)
+        msgs.info('Excluding the following slits:')
+        for islit in _exclude_slits:
+            if islit in slits0.slitord_id[good_slitindx]:
+                msgs.info(f'Slit {islit}')
+                exclude_slitindx = np.append(exclude_slitindx,
+                                             np.where(slits0.slitord_id[good_slitindx] == islit)[0][0])
+        # these are the good slit index excluding the slits that are selected by the user
+        return np.delete(good_slitindx, exclude_slitindx)
 
     def optimal_weights(self, slitorderid, objid, const_weights=False):
         """
@@ -421,7 +487,7 @@ class CoAdd2D:
         waves, fluxes, ivars, gpms = [], [], [], []
 
         for iexp, sobjs in enumerate(self.stack_dict['specobjs_list']):
-            ithis = sobjs.slitorder_objid_indices(slitorderid, objid[iexp])
+            ithis = sobjs.slitorder_objid_indices(slitorderid, objid[iexp], toler=self.par['coadd2d']['spat_toler'])
             if not np.any(ithis):
                 msgs.error('Slit/order or OBJID provided not valid. Optimal weights cannot be determined.')
             # check if OPT_COUNTS is available
@@ -448,7 +514,7 @@ class CoAdd2D:
         rms_sn, weights = coadd.sn_weights(fluxes, ivars, gpms, self.sn_smooth_npix, const_weights=const_weights)
         return rms_sn, weights
 
-    def coadd(self, only_slits=None, interp_dspat=True):
+    def coadd(self, interp_dspat=True):
         """
         Construct a 2d co-add of a stack of PypeIt spec2d reduction outputs.
         This method calls loops over slits/orders and performs the 2d-coadd by
@@ -457,10 +523,6 @@ class CoAdd2D:
 
         Parameters
         ----------
-        only_slits : list, optional
-           List of slits to operate on. Not currently supported, i.e. the code
-           can currently only stack everything because the slit/reduction
-           bitmask checking is not yet implemented. Default = None
         interp_dspat : bool, optional
            Interpolate in the spatial coordinate image to faciliate running
            through core.extract.local_skysub_extract.  Default=True
@@ -473,27 +535,32 @@ class CoAdd2D:
             # TODO Make this a PypeIt object, with data model yada-yada.
 
         """
-        # get slit index that indicates which slits are good for coadding
-        self.good_slits = self.good_slitindx(only_slits=only_slits)
-        # get the number of slits that are going to be coadded
-        self.nslits_coadded = self.good_slits.size
 
         coadd_list = []
         for slit_idx in self.good_slits:
-            msgs.info('Performing 2d coadd for slit: {:d}/{:d}'.format(slit_idx, self.nslits_single - 1))
-            ref_trace_stack = self.reference_trace_stack(slit_idx, offsets=self.offsets,
-                                                         objid=self.objid_bri)
+            msgs.info(f'Performing 2D coadd for slit {self.spat_ids[slit_idx]} ({slit_idx + 1}/{self.nslits_single})')
 
-            thismask_stack = [np.abs(slitmask - self.stack_dict['slits_list'][0].spat_id[slit_idx]) <= self.par['coadd2d']['spat_toler'] for slitmask in self.stack_dict['slitmask_stack']]
+            # mask identifying the current slit in each exposure
+            thismask_stack = [np.abs(slitmask - self.spat_ids[slit_idx]) <= self.par['coadd2d']['spat_toler']
+                              for slitmask in self.stack_dict['slitmask_stack']]
+
+            # check if the slit is found in every exposure
+            if not np.all([np.any(thismask) for thismask in thismask_stack]):
+                msgs.warn(f'Slit {self.spat_ids[slit_idx]} was not found in every exposures. '
+                          f'2D coadd cannot be performed on this slit. Try increasing the parameter spat_toler')
+                continue
+
+            # reference trace
+            ref_trace_stack = self.reference_trace_stack(slit_idx, offsets=self.offsets, objid=self.objid_bri)
             # maskdef info
             maskdef_dict = self.get_maskdef_dict(slit_idx, ref_trace_stack)
 
             # weights
-            ## JFH I do not understand why this is here and am removing it.
-            #if not isinstance(self.use_weights, str):
-            #    weights = self.use_weights[slit_idx]
-            #else:
-            #    weights = self.use_weights
+            # if this is echelle data and the parset 'weights' is set to 'auto',
+            # then the weights are computed per order, i.e., every order has a
+            # different set of weights in each exposure (len(self.use_weights[slit_idx]) = nexp)
+            _weights = self.use_weights[slit_idx] if self.pypeline == 'Echelle' and self.weights == 'auto' else self.use_weights
+            # TODO: Create a method here in the child clases? It is not great to do pypeline specific things in the parent
 
             # Perform the 2d coadd
             # NOTE: mask_stack is a gpm, and this is called inmask_stack in
@@ -507,8 +574,11 @@ class CoAdd2D:
                                                self.stack_dict['waveimg_stack'],
                                                self.wave_grid, self.spat_samp_fact,
                                                maskdef_dict=maskdef_dict,
-                                               weights=self.use_weights, interp_dspat=interp_dspat)
+                                               weights=_weights, interp_dspat=interp_dspat)
             coadd_list.append(coadd_dict)
+
+        if len(coadd_list) == 0:
+            msgs.error("All the slits were missing in one or more exposures. 2D coadd cannot be performed")
 
         return coadd_list
 
@@ -615,7 +685,7 @@ class CoAdd2D:
             if None not in all_maskdef_ids:
                 maskdef_id[islit] = coadd_dict['maskdef_id']
                 maskdef_objpos[islit] = coadd_dict['maskdef_objpos']
-                maskdef_slitcen[:, islit] = np.full(nspec_pseudo, coadd_dict['maskdef_slitcen'])
+                maskdef_slitcen[:, islit] = np.full(nspec_pseudo, coadd_dict['maskdef_slitcen'] + slit_left[:,islit])
                 if coadd_dict['maskdef_designtab'] is not None:
                     maskdef_designtab = vstack([maskdef_designtab, coadd_dict['maskdef_designtab']])
 
@@ -822,7 +892,7 @@ class CoAdd2D:
             :obj:`list`: A list of reference traces for the 2d coadding that
             have been offset.
         """
-        return [slits.center[:,slitid] - offsets[iexp] 
+        return [slits.center[:,slitid] - offsets[iexp]
                     for iexp, slits in enumerate(self.stack_dict['slits_list'])]
 #        ref_trace_stack = []
 #        for iexp, slits in enumerate(self.stack_dict['slits_list']):
@@ -1090,8 +1160,6 @@ class CoAdd2D:
                 msgs.info('Using uniform weights')
             # use uniform weights
             self.use_weights = (np.ones(self.nexp) / float(self.nexp)).tolist()
-            #self.use_weights = 'uniform'
-
 
         # 3) Bright object exists and parset `weights` is equal to 'auto'
         elif (self.objid_bri is not None) and (weights == 'auto'):
@@ -1165,13 +1233,15 @@ class MultiSlitCoAdd2D(CoAdd2D):
 
     """
     def __init__(self, spec2d_files, spectrograph, par, det=1, offsets=None, weights='auto',
+                 only_slits=None, exclude_slits=None,
                  spec_samp_fact=1.0, spat_samp_fact=1.0, sn_smooth_npix=None,
                  bkg_redux=False, find_negative=False, show=False, show_peaks=False, debug_offsets=False, debug=False):
         super().__init__(spec2d_files, spectrograph, det=det, offsets=offsets, weights=weights,
-                                               spec_samp_fact=spec_samp_fact, spat_samp_fact=spat_samp_fact,
-                                               sn_smooth_npix=sn_smooth_npix, bkg_redux=bkg_redux, find_negative=find_negative, par=par,
-                                        show=show, show_peaks=show_peaks, debug_offsets=debug_offsets,
-                                        debug=debug)
+                         only_slits=only_slits, exclude_slits=exclude_slits,
+                         spec_samp_fact=spec_samp_fact, spat_samp_fact=spat_samp_fact,
+                         sn_smooth_npix=sn_smooth_npix, bkg_redux=bkg_redux, find_negative=find_negative, par=par,
+                         show=show, show_peaks=show_peaks, debug_offsets=debug_offsets,
+                         debug=debug)
 
         # maskdef offset
         self.maskdef_offset = np.array([slits.maskdef_offset for slits in self.stack_dict['slits_list']])
@@ -1189,13 +1259,14 @@ class MultiSlitCoAdd2D(CoAdd2D):
                 # does it exists?
                 user_obj_exist = []
                 for sobjs in self.stack_dict['specobjs_list']:
-                    user_obj_exist.append(np.any(sobjs.slitorder_objid_indices(user_slit, user_objid)))
+                    user_obj_exist.append(np.any(sobjs.slitorder_objid_indices(user_slit, user_objid,
+                                                                               toler=self.par['coadd2d']['spat_toler'])))
                 if not np.all(user_obj_exist):
                     msgs.error('Object provided through `user_obj` does not exist in all the exposures.')
 
         # find if there is a bright object we could use
         if len(self.stack_dict['specobjs_list']) > 0 and self.par['coadd2d']['user_obj'] is not None:
-            _slitidx_bri = np.where(self.spat_ids == user_slit)[0][0]
+            _slitidx_bri = np.where(np.abs(self.spat_ids - user_slit) <= self.par['coadd2d']['spat_toler'])[0][0]
             self.objid_bri, self.slitidx_bri, self.spatid_bri, self.snr_bar_bri = \
                 np.repeat(user_objid, self.nexp), _slitidx_bri, user_slit, None
         elif len(self.stack_dict['specobjs_list']) > 0 and (offsets == 'auto' or weights == 'auto'):
@@ -1240,7 +1311,7 @@ class MultiSlitCoAdd2D(CoAdd2D):
             thismask_stack = [np.abs(slitmask - self.spatid_bri) <= self.par['coadd2d']['spat_toler'] for slitmask in self.stack_dict['slitmask_stack']]
 
             # TODO Need to think abbout whether we have multiple tslits_dict for each exposure or a single one
-            trace_stack_bri = [slits.center[:, self.slitidx_bri] 
+            trace_stack_bri = [slits.center[:, self.slitidx_bri]
                                     for slits in self.stack_dict['slits_list']]
 #            trace_stack_bri = []
 #            for slits in self.stack_dict['slits_list']:
@@ -1345,28 +1416,24 @@ class MultiSlitCoAdd2D(CoAdd2D):
         slit_snr_max = np.zeros((nslits, nexp), dtype=float)
         bpm = np.ones(slit_snr_max.shape, dtype=bool)
         objid_max = np.zeros((nslits, nexp), dtype=int)
-        # Loop over each exposure, slit, find the brighest object on that slit for every exposure
+        # Loop over each exposure, slit, find the brightest object on that slit for every exposure
         for iexp, sobjs in enumerate(specobjs_list):
             msgs.info("Working on exposure {}".format(iexp))
-            #nspec_now = self.nspec_array[iexp]
             for islit, spat_id in enumerate(spat_ids):
                 ithis = np.abs(sobjs.SLITID - spat_id) <= self.par['coadd2d']['spat_toler']
-                #nobj_slit = np.sum(ithis)
                 if np.any(ithis):
                     objid_this = sobjs[ithis].OBJID
-                    waves, fluxes, ivars, gpms = [], [], [], []
+                    fluxes, ivars, gpms = [], [], []
                     for iobj, spec in enumerate(sobjs[ithis]):
                         # check if OPT_COUNTS is available
                         if spec.has_opt_ext():
-                            wave_iobj, flux_iobj, ivar_iobj, gpm_iobj = spec.get_opt_ext()
-                            waves.append(wave_iobj)
+                            _, flux_iobj, ivar_iobj, gpm_iobj = spec.get_opt_ext()
                             fluxes.append(flux_iobj)
                             ivars.append(ivar_iobj)
                             gpms.append(gpm_iobj)
                         # check if BOX_COUNTS is available
                         elif spec.has_box_ext():
-                            wave_iobj, flux_iobj, ivar_iobj, gpm_iobj = spec.get_box_ext()
-                            waves.append(wave_iobj)
+                            _, flux_iobj, ivar_iobj, gpm_iobj = spec.get_box_ext()
                             fluxes.append(flux_iobj)
                             ivars.append(ivar_iobj)
                             gpms.append(gpm_iobj)
@@ -1378,17 +1445,24 @@ class MultiSlitCoAdd2D(CoAdd2D):
                             msgs.warn(f'Optimal and Boxcar extraction not available for obj {spec.OBJID} in slit {spat_id}.')
                             #remove_indx.append(iobj)
                     # if there are objects on this slit left, we can proceed with computing rms_sn
-                    if len(waves) > 0:
+                    if len(fluxes) > 0:
                         rms_sn, weights = coadd.sn_weights(fluxes, ivars, gpms, const_weights=True)
                         imax = np.argmax(rms_sn)
                         slit_snr_max[islit, iexp] = rms_sn[imax]
                         objid_max[islit, iexp] = objid_this[imax]
                         bpm[islit, iexp] = False
 
+        # If a slit has bpm = True for some exposures and not for others, set bpm = True for all exposures
+        # Find the rows where any of the bpm values are True
+        bpm_true_idx = np.array([np.any(b) for b in bpm])
+        if np.any(bpm_true_idx):
+            # Flag all exposures in those rows
+            bpm[bpm_true_idx, :] = True
+
         # Find the highest snr object among all the slits
         if np.all(bpm):
             msgs.warn('You do not appear to have a unique reference object that was traced as the highest S/N '
-                      'ratio on the same slit of every exposure')
+                      'ratio on the same slit of every exposure. Try increasing the parameter `spat_toler`')
             return None, None, None, None
         else:
             # mask the bpm
@@ -1398,6 +1472,7 @@ class MultiSlitCoAdd2D(CoAdd2D):
             snr_bar_mean = slit_snr[slitid]
             snr_bar = slit_snr_max[slitid, :]
             objid = objid_max[slitid, :]
+
         return objid, slitid, spat_ids[slitid], snr_bar
 
     # TODO add an option here to actually use the reference trace for cases where they are on the same slit and it is
@@ -1438,9 +1513,9 @@ class MultiSlitCoAdd2D(CoAdd2D):
                 self.stack_dict['slits_list'][0].maskdef_objpos is not None and \
                 self.stack_dict['maskdef_designtab_list'][0] is not None:
             # maskdef_designtab info for only this slit
-            this_idx = self.stack_dict['maskdef_designtab_list'][0]['SPAT_ID'] == self.stack_dict['slits_list'][0].spat_id[slit_idx]
+            this_idx = self.stack_dict['maskdef_designtab_list'][0]['SPAT_ID'] == self.spat_ids[slit_idx]
             this_maskdef_designtab = self.stack_dict['maskdef_designtab_list'][0][this_idx]
-            # remove columns that that are irrelevant in the coadd2d frames
+            # remove columns that are irrelevant in the coadd2d frames
             this_maskdef_designtab.remove_columns(['TRACEID', 'TRACESROW', 'TRACELPIX', 'TRACERPIX',
                                                    'SLITLMASKDEF', 'SLITRMASKDEF'])
             this_maskdef_designtab.meta['MASKRMSL'] = 0.
@@ -1449,36 +1524,36 @@ class MultiSlitCoAdd2D(CoAdd2D):
             # maskdef_id for this slit
             imaskdef_id = self.stack_dict['slits_list'][0].maskdef_id[slit_idx]
 
-            # maskdef_slitcenters. This trace the slit center along the spectral direction.
-            # But here we take only the value at the mid point
+            # maskdef_slitcen (slit center along the spectral direction) and
+            # maskdef_objpos (expected position of the target, as distance from left slit edge) for this slit
 
-            maskdef_slitcen_pixpos = self.stack_dict['slits_list'][0].maskdef_slitcen[self.nspec_array[0]//2, slit_idx] + self.maskdef_offset
-            # binned maskdef_slitcenters position with respect to the center of the slit in ref_trace_stack
-            # this value should be the same for each exposure, but in case there are differences we take the mean value
-
+            # These are the binned maskdef_slitcen positions w.r.t. the center of the slit in ref_trace_stack
             slit_cen_dspat_vec = np.zeros(self.nexp)
-            for iexp, (ref_trace, maskdef_slitcen) in enumerate(zip(ref_trace_stack, maskdef_slitcen_pixpos)):
+            # These are the binned maskdef_objpos positions w.r.t. the center of the slit in ref_trace_stack
+            objpos_dspat_vec = np.zeros(self.nexp)
+            for iexp in range(self.nexp):
+                # get maskdef_slitcen
+                maskdef_slitcen_pixpos = \
+                    self.stack_dict['slits_list'][iexp].maskdef_slitcen[self.nspec_array[0]//2, slit_idx] \
+                    + self.maskdef_offset[iexp]
+
+                # get maskdef_objpos
+                # find left edge
+                slits_left, _, _ = \
+                    self.stack_dict['slits_list'][iexp].select_edges(flexure=self.stack_dict['spat_flexure_list'][iexp])
+                # targeted object spat pix
+                maskdef_obj_pixpos = \
+                    self.stack_dict['slits_list'][iexp].maskdef_objpos[slit_idx] + self.maskdef_offset[iexp] \
+                    + slits_left[slits_left.shape[0]//2, slit_idx]
+
+                # change reference system
+                ref_trace = ref_trace_stack[iexp]
                 nspec_this = ref_trace.shape[0]
-                slit_cen_dspat_vec[iexp] = (maskdef_slitcen - ref_trace[nspec_this//2])/self.spat_samp_fact
+                slit_cen_dspat_vec[iexp] = (maskdef_slitcen_pixpos - ref_trace[nspec_this // 2]) / self.spat_samp_fact
+
+                objpos_dspat_vec[iexp] = (maskdef_obj_pixpos - ref_trace[nspec_this // 2]) / self.spat_samp_fact
 
             imaskdef_slitcen_dspat = np.mean(slit_cen_dspat_vec)
-
-            # expected position of the targeted object from slitmask design (as distance from left slit edge)
-            imaskdef_objpos = self.stack_dict['slits_list'][0].maskdef_objpos[slit_idx]
-
-            # find left edge
-            slits_left, _, _ = self.stack_dict['slits_list'][0].select_edges(flexure=self.stack_dict['spat_flexure_list'][0])
-            # targeted object spat pix
-            nspec_this = slits_left.shape[0]
-            maskdef_obj_pixpos = imaskdef_objpos + self.maskdef_offset + slits_left[nspec_this//2, slit_idx]
-            # binned expected object position with respect to the center of the slit in ref_trace_stack
-            # this value should be the same for each exposure, but in case there are differences we take the mean value
-
-            objpos_dspat_vec = np.zeros(self.nexp)
-            for iexp, (ref_trace, maskdef_obj) in enumerate(zip(ref_trace_stack, maskdef_obj_pixpos)):
-                nspec_this = ref_trace.shape[0]
-                objpos_dspat_vec[iexp] = (maskdef_obj - ref_trace[nspec_this//2])/self.spat_samp_fact
-
             imaskdef_objpos_dspat = np.mean(objpos_dspat_vec)
 
         else:
@@ -1510,13 +1585,15 @@ class EchelleCoAdd2D(CoAdd2D):
 
     """
     def __init__(self, spec2d_files, spectrograph, par, det=1, offsets=None, weights='auto',
+                 only_slits=None, exclude_slits=None,
                  spec_samp_fact=1.0, spat_samp_fact=1.0, sn_smooth_npix=None,
                  bkg_redux=False, find_negative=False, show=False, show_peaks=False, debug_offsets=False, debug=False):
         super().__init__(spec2d_files, spectrograph, det=det, offsets=offsets, weights=weights,
-                                             spec_samp_fact=spec_samp_fact, spat_samp_fact=spat_samp_fact,
-                                             sn_smooth_npix=sn_smooth_npix, bkg_redux=bkg_redux, find_negative=find_negative,
-                                             par=par, show=show, show_peaks=show_peaks, debug_offsets=debug_offsets,
-                                             debug=debug)
+                         only_slits=only_slits, exclude_slits=exclude_slits,
+                         spec_samp_fact=spec_samp_fact, spat_samp_fact=spat_samp_fact,
+                         sn_smooth_npix=sn_smooth_npix, bkg_redux=bkg_redux, find_negative=find_negative,
+                         par=par, show=show, show_peaks=show_peaks, debug_offsets=debug_offsets,
+                         debug=debug)
 
         # Default wave_method for Echelle is log10
         wave_method = 'log10' if self.par['coadd2d']['wave_method'] is None else self.par['coadd2d']['wave_method']
@@ -1594,12 +1671,10 @@ class EchelleCoAdd2D(CoAdd2D):
         if (self.objid_bri is not None) and (weights == 'auto'):
             # computing a list of weights for all the slitord_ids that we than parse in coadd
             slitord_ids = self.stack_dict['slits_list'][0].slitord_id
-            use_weights = []
-            for id in slitord_ids:
-                _, iweights = self.optimal_weights(id, self.objid_bri)
-                use_weights.append(iweights)
-            # Commenting this out since self.use_weights needs to be a list or a string
-            #self.use_weights = np.array(use_weights)
+            self.use_weights = []
+            for idx in slitord_ids:
+                _, iweights = self.optimal_weights(idx, self.objid_bri)
+                self.use_weights.append(iweights)
             if self.par['coadd2d']['user_obj'] is not None:
                 msgs.info('Weights computed using a unique reference object provided by the user')
             else:
@@ -1628,7 +1703,6 @@ class EchelleCoAdd2D(CoAdd2D):
 
         objid = np.zeros(nexp, dtype=int)
         snr_bar = np.zeros(nexp)
-        # norders = specobjs_list[0].ech_orderindx.max() + 1
         for iexp, sobjs in enumerate(specobjs_list):
             msgs.info("Working on exposure {}".format(iexp))
             uni_objid = np.unique(sobjs.ECH_OBJID)
@@ -1637,24 +1711,31 @@ class EchelleCoAdd2D(CoAdd2D):
             bpm = np.ones((nslits, nobjs), dtype=bool)
             for iord in range(nslits):
                 for iobj in range(nobjs):
+                    flux = None
                     ind = (sobjs.ECH_ORDERINDX == iord) & (sobjs.ECH_OBJID == uni_objid[iobj])
                     # check if OPT_COUNTS is available
                     if sobjs[ind][0].has_opt_ext():
-                        wave, flux, ivar, mask = sobjs[ind][0].get_opt_ext()
+                        _, flux, ivar, mask = sobjs[ind][0].get_opt_ext()
                     # check if BOX_COUNTS is available
                     elif sobjs[ind][0].has_box_ext():
-                        wave, flux, ivar, mask = sobjs[ind][0].get_box_ext()
+                        _, flux, ivar, mask = sobjs[ind][0].get_box_ext()
                         msgs.warn(f'Optimal extraction not available for object {sobjs[ind][0].ECH_OBJID} '
                                   f'in order {sobjs[ind][0].ECH_ORDER}. Using box extraction.')
                     else:
-                        flux = None
                         msgs.warn(f'Optimal and Boxcar extraction not available for '
                                   f'object {sobjs[ind][0].ECH_OBJID} in order {sobjs[ind][0].ECH_ORDER}.')
                         continue
                     if flux is not None:
-                        rms_sn, weights = coadd.sn_weights(wave, flux, ivar, mask, const_weights=True)
+                        rms_sn, weights = coadd.sn_weights([flux], [ivar], [mask], const_weights=True)
                         order_snr[iord, iobj] = rms_sn
                         bpm[iord, iobj] = False
+
+            # If there are orders that have bpm = True for some objs and not for others, set bpm = True for all objs
+            # Find the rows where any of the bpm values are True
+            bpm_true_idx = np.array([np.any(b) for b in bpm])
+            if np.any(bpm_true_idx):
+                # Flag all objs in those rows
+                bpm[bpm_true_idx, :] = True
 
             # Compute the average SNR and find the brightest object
             if not np.all(bpm):