+++ pypeit/spectrographs/shane_kast.py
@@ -48,7 +48,7 @@ class ShaneKastSpectrograph(spectrograph.Spectrograph):
         # Always correct for flexure, starting with default parameters
         par['flexure']['spec_method'] = 'boxcar'
         # Set the default exposure time ranges for the frame typing
-        par['calibrations']['biasframe']['exprng'] = [None, 1]
+        par['calibrations']['biasframe']['exprng'] = [None, 0.001]
         par['calibrations']['darkframe']['exprng'] = [999999, None]     # No dark frames
         par['calibrations']['pinholeframe']['exprng'] = [999999, None]  # No pinhole frames
         par['calibrations']['pixelflatframe']['exprng'] = [0, None]